feat(jss): add lib/jss to augment reacat-jss in order to use a more elegannt and type strict api
This change adds 2 new functions "createStyles" and "createUseStyles" that are meant to be used
to define styles.

"createStyles" is a HOC component and it's the only way a class based component can be styled, but
it also works for Function Components when it's. I'd say it's not advised to though, but we'll see :)
"createUseStyles" is a Hook and this works only for function Based components.