feat(lib/jss): improve the API a bit more
Add ability to use it like this:

const styled = createStyles({
  container: {
    position: 'relative',
    bottom: '2',
    left: '80%',
  },
});

export const SplashScreenBoardWithButtons = styled<Props>(({ classes }) => (
  <div classNames={classes.container}/>
);

--

As you can see this is much more terse, as it only makes use of 1 function: createStyles
which also bind the Props.classes with the correct defined classes ("container").

The only problem with this approach is that the styles need to be defined before the
component, not after.