feat(room provider): add Provider to keep the Socket and RTC Conns in sync. Exposes new Peer & Room
This component is is the bridge from the lowerlevel Socket Connections and PeersProvider and combines
them into one, thus simplifying the API for a Room Container.

For now, only the Peers that have an active RTC Connections are passed further. This means that if
the Socket Somehow gets out of sync either by having more connections or less than RTC those won't be
reflected. This is probably not the best scenario but it suffices the needs for now as the only important
connections are the RTC ones and as long as those are active there shouldn't be any discrepence.

There is some thought to be put in what happens when a Peer drops the socket connection, should she still
be part of the room, (and for now she is) but maybe later on dropping the socket connection will not be
an exception but part of the design – since we might not won't the Peers to be always connected to
the server, which could be very hard on the resources.

We shall see, what the futture needs hold :)