fix rtc conn not reove don disconnect. Fix onMessageSent not getting called
The problem with the 1st one is pretty simple, yet a really annoying bug, as
the rtc connections weren't getting updated. This solution is just a patch.
The proper solution is to maintain only rtc connection state, while all the
others are either derivates or use this one directly.

The second one issue was related to the Peer.onPeerMsgSent() not getting
triggered if there were any bad sends (due to a dropped connection or smtg)
while the intended design was the opposite – to trigger it if there was
at least one good send