fix(services/peers): fix the issue of local streaming not being add to the track only sometimes
The problem here was not with RTC but with the dact that I was using a singleton local stream
(always returning the active one), which of course doesn't work with streaming it to multiple peers.
I got the solution by testing it in Firefox, which had a better error message and made it more clear
what the cause it.