fix(peerprovider): fix issue with Streaming box display my name. Fix my stream stopping on peer exit
The 1st one was b/c I wasn't using the correct metadata.

The 2nd because I was adding my new stream (for each remote peer) to the store, thus replacin the old one,
which was just local. Once the Remote Peer existed the room and thus I stopped the tracks from my
designated local, that would stop. My local stream is and should be seperate from the remote peer streams
which is just temporary