feat(guest authentication): ensure a stale guest user is not used anymore.
A Stale Guest User means a User that was created by a previous Server Instance and thus
a DB instance (since the DB gets flushed). In fact the DB instance is what matters the most.
so in the future I should look into that.

This change is twofolded:
 - on the client, ensure upon authenticating a new User (guest o registered) the current one gets unset
 - on the server, verify that the sid (server id) that was used to create this Guest is the same, otherwise
just generate a new Guest