feat(peerprovider): a set of changes that optimizes for SEO and PeerJS usage
The changes are:
1. SEO Optimization
  The content inside the <PeerProvider> doesn't wait until a connection to the server
is made before being displayed. This in addition with prerendering improves SEO
dramatically as well User Experience in general making the load time much shorter

2. PeerJS optimization
  The PeerProvider is now smart enough to connect to the Signaling Server only when inside
a Room, and disconnects once the room is gone.

A Peer is:
 - An Authenticated User that has an active Socket Connection and is UserIdentified!

A User can be a Peer eventhough it's not connected to the Peer Signaling Server anymore!
A Peer is a domain label for chessroulette only and not RTC or PeerJS