refactor(avstreaming): change the getUserMedia management. Use cloning of an active stream!
This change is needed b/c calling getUserMedia when one is already active results in a freeze
on Safari, and I also believe it's worse performance wise. What I do instead is cloning the
active one.

In terms of destroying the userMedia, that happens when there is none used anymore, and it's
important to destroy it so the camera/mic doesn't remain active when not displayed. This could
be a problem with battery on mobile and als privacy so we don't wont to keep user media active
when not explicitly used!