refactor(room bouncer): some big changes on the Permission UX
First big change is taht now the room bouncer state gets partially
persisted, and the permissions get automatically checked if already have
gone once. This works perfectly in chromei and firefox as it doesn't asks the user
anymore, but not on Safari.

Secondly, there is no need reconfirm joining a room after having joined it once
before. This works great with rereshes. Again Chrome works perfectly as expected,
but Safari asks for Camera/Mic Permissions again. Not the worst as it's only
one step.