fix(room): continue the Refactorng. start using useJoinRoom to build each Activity
In addition, I changed the way the participant, member, peer depend on each other, moving from a
overwritten flat approach to a nested approach. Stil not 100% if it's better but it does make more
sense as now everyhting is namespaced and still accessible. In addition I believe it should work
correctly and optimize space but not 100%. Need to run some more tests as well as do some more
research