refactor(usejoinedroom): optimize it to not trigger a rerender if state value hasn't changed
This is actually an optimization for react renders. Since the server was sending a new update each time
this simply ensure that the state hasn't updated!

In the future I could do have an identifier (checksum) and simply check if thos have changed on the payloads,
if not don't even bother :)