refactor(chessgamehistoryprovider): add a displayed obj on the context that holds an atomic snapshot
This currently encapsulated::
- history
- pgn
- index
- fen

Adnt it's guaranteed that they are all in sync, meaning if one changes it's ok to check the fen or pgn
for updates.

In additoin the ChessGame was changed to use the new api, and not call the useGameActions.onMove by itself
anymore, but to only call a passed prop that happens to come from the GameHistoryConsumer.onAddMove.

This way the GameHistoryProvider can setup the display props before the server returns a result, which
ensures the client is fast.